requirements.txt
openai
langchain
langchain-openai
langchain-ollama
python-dotenv
pypdf
langchain-community
azure-core
azure-search-documents
ollama
gpt4all
certifi
llama-cpp-python
chromadb